♻️move experimental features to tools
if it was domain logic, we should not use it on tools code
