+++ packages/core/src/browser/pageExitObservable.spec.ts
@@ -1,5 +1,5 @@
 import { createNewEvent, restorePageVisibility, setPageVisibility } from '../../test'
-import { resetExperimentalFeatures, addExperimentalFeatures, ExperimentalFeature } from '../domain/configuration'
+import { resetExperimentalFeatures, addExperimentalFeatures, ExperimentalFeature } from '../tools/experimentalFeatures'
 import type { Subscription } from '../tools/observable'
 import type { PageExitEvent } from './pageExitObservable'
 import { PageExitReason, createPageExitObservable } from './pageExitObservable'
