+++ packages/core/src/domain/configuration/configuration.spec.ts
@@ -1,8 +1,12 @@
 import type { RumEvent } from '../../../../rum-core/src'
 import { display } from '../../tools/display'
+import {
+  ExperimentalFeature,
+  isExperimentalFeatureEnabled,
+  resetExperimentalFeatures,
+} from '../../tools/experimentalFeatures'
 import type { InitConfiguration } from './configuration'
 import { validateAndBuildConfiguration } from './configuration'
-import { ExperimentalFeature, isExperimentalFeatureEnabled, resetExperimentalFeatures } from './experimentalFeatures'
 
 describe('validateAndBuildConfiguration', () => {
   const clientToken = 'some_client_token'
