+++ packages/core/src/domain/configuration/endpointBuilder.ts
@@ -2,9 +2,9 @@ import type { RetryInfo, FlushReason } from '../../transport'
 import { timeStampNow } from '../../tools/timeUtils'
 import { normalizeUrl } from '../../tools/urlPolyfill'
 import { generateUUID } from '../../tools/utils'
+import { ExperimentalFeature, isExperimentalFeatureEnabled } from '../../tools/experimentalFeatures'
 import type { InitConfiguration } from './configuration'
 import { INTAKE_SITE_AP1, INTAKE_SITE_US1 } from './intakeSites'
-import { ExperimentalFeature, isExperimentalFeatureEnabled } from './experimentalFeatures'
 
 // replaced at build time
 declare const __BUILD_ENV__SDK_VERSION__: string
