+++ packages/core/src/domain/configuration/index.ts
@@ -7,11 +7,4 @@ export {
   serializeConfiguration,
 } from './configuration'
 export { createEndpointBuilder, EndpointBuilder, EndpointType } from './endpointBuilder'
-export {
-  isExperimentalFeatureEnabled,
-  addExperimentalFeatures,
-  resetExperimentalFeatures,
-  getExperimentalFeatures,
-  ExperimentalFeature,
-} from './experimentalFeatures'
 export * from './intakeSites'
