+++ packages/core/src/domain/console/consoleObservable.ts
@@ -5,7 +5,7 @@ import { find, jsonStringify } from '../../tools/utils'
 import { ConsoleApiName } from '../../tools/display'
 import { callMonitored } from '../../tools/monitor'
 import { sanitize } from '../../tools/sanitize'
-import { ExperimentalFeature, isExperimentalFeatureEnabled } from '../configuration'
+import { ExperimentalFeature, isExperimentalFeatureEnabled } from '../../tools/experimentalFeatures'
 
 export interface ConsoleLog {
   message: string
