+++ packages/core/src/tools/limitModification.ts
@@ -1,4 +1,4 @@
-import { ExperimentalFeature, isExperimentalFeatureEnabled } from '../domain/configuration'
+import { ExperimentalFeature, isExperimentalFeatureEnabled } from './experimentalFeatures'
 import type { Context } from './context'
 import { sanitize } from './sanitize'
 import { deepClone, getType } from './utils'
